+++ build-logic/gobley-gradle-cargo/src/main/kotlin/CargoPlugin.kt
@@ -25,6 +25,10 @@ import gobley.gradle.cargo.tasks.CargoTask
 import gobley.gradle.cargo.tasks.RustUpTargetAddTask
 import gobley.gradle.cargo.tasks.RustUpTask
 import gobley.gradle.cargo.utils.register
+import gobley.gradle.kotlin.GobleyKotlinAndroidExtensionDelegate
+import gobley.gradle.kotlin.GobleyKotlinExtensionDelegate
+import gobley.gradle.kotlin.GobleyKotlinJvmExtensionDelegate
+import gobley.gradle.kotlin.GobleyKotlinMultiplatformExtensionDelegate
 import gobley.gradle.rust.CrateType
 import gobley.gradle.rust.targets.RustAndroidTarget
 import gobley.gradle.rust.targets.RustJvmTarget
@@ -47,17 +51,16 @@ import org.gradle.api.tasks.Delete
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.jvm.tasks.Jar
 import org.gradle.kotlin.dsl.create
-import org.gradle.kotlin.dsl.getByType
 import org.gradle.kotlin.dsl.listProperty
 import org.gradle.kotlin.dsl.named
 import org.gradle.kotlin.dsl.register
 import org.gradle.kotlin.dsl.withType
 import org.gradle.language.jvm.tasks.ProcessResources
-import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 import org.jetbrains.kotlin.gradle.plugin.KotlinPlatformType
 import org.jetbrains.kotlin.gradle.plugin.KotlinTarget
 import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinAndroidTarget
 import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinNativeTarget
+import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinWithJavaTarget
 import org.jetbrains.kotlin.gradle.targets.jvm.KotlinJvmTarget
 import java.io.File
 
@@ -67,7 +70,9 @@ class CargoPlugin : Plugin<Project> {
     }
 
     private lateinit var cargoExtension: CargoExtension
-    private lateinit var kotlinMultiplatformExtension: KotlinMultiplatformExtension
+
+    @OptIn(InternalGobleyGradleApi::class)
+    private lateinit var kotlinExtensionDelegate: GobleyKotlinExtensionDelegate
     private lateinit var androidDelegate: CargoPluginAndroidDelegate
 
     override fun apply(target: Project) {
@@ -114,8 +119,16 @@ class CargoPlugin : Plugin<Project> {
     @OptIn(InternalGobleyGradleApi::class)
     private fun Project.watchPluginChanges() {
         plugins.withId(PluginIds.KOTLIN_MULTIPLATFORM) {
-            kotlinMultiplatformExtension = extensions.getByType()
-            kotlinMultiplatformExtension.targets.configureEach { planBuilds() }
+            kotlinExtensionDelegate = GobleyKotlinMultiplatformExtensionDelegate(project)
+            kotlinExtensionDelegate.targets.configureEach { planBuilds() }
+        }
+        plugins.withId(PluginIds.KOTLIN_ANDROID) {
+            kotlinExtensionDelegate = GobleyKotlinAndroidExtensionDelegate(project)
+            kotlinExtensionDelegate.targets.configureEach { planBuilds() }
+        }
+        plugins.withId(PluginIds.KOTLIN_JVM) {
+            kotlinExtensionDelegate = GobleyKotlinJvmExtensionDelegate(project)
+            kotlinExtensionDelegate.targets.configureEach { planBuilds() }
         }
 
         val androidPluginAction = Action<Plugin<*>> {
@@ -139,8 +152,18 @@ class CargoPlugin : Plugin<Project> {
         @OptIn(InternalGobleyGradleApi::class)
         PluginUtils.ensurePluginIsApplied(
             this,
-            "Kotlin Multiplatform",
-            PluginIds.KOTLIN_MULTIPLATFORM
+            PluginUtils.PluginInfo(
+                "Kotlin Multiplatform",
+                PluginIds.KOTLIN_MULTIPLATFORM
+            ),
+            PluginUtils.PluginInfo(
+                "Kotlin Android",
+                PluginIds.KOTLIN_ANDROID,
+            ),
+            PluginUtils.PluginInfo(
+                "Kotlin JVM",
+                PluginIds.KOTLIN_JVM,
+            ),
         )
     }
 
@@ -152,28 +175,29 @@ class CargoPlugin : Plugin<Project> {
 
     private fun KotlinTarget.requiredRustTargets(): List<RustTarget> {
         return when (this) {
-            is KotlinJvmTarget -> GobleyHost.current.platform.supportedTargets.filterIsInstance<RustJvmTarget>()
+            is KotlinJvmTarget, is KotlinWithJavaTarget<*, *> -> GobleyHost.current.platform.supportedTargets.filterIsInstance<RustJvmTarget>()
             is KotlinAndroidTarget -> RustAndroidTarget.values().toList()
             is KotlinNativeTarget -> listOf(RustTarget(konanTarget))
             else -> listOf()
         }
     }
 
+    @OptIn(InternalGobleyGradleApi::class)
     private fun Project.checkKotlinTargets() {
         val hasJsTargets =
-            kotlinMultiplatformExtension.targets.any { it.platformType == KotlinPlatformType.js }
+            kotlinExtensionDelegate.targets.any { it.platformType == KotlinPlatformType.js }
         if (hasJsTargets) {
             project.logger.warn("JS targets are added, but UniFFI KMP bindings does not support JS targets yet.")
         }
 
         val hasWasmTargets =
-            kotlinMultiplatformExtension.targets.any { it.platformType == KotlinPlatformType.wasm }
+            kotlinExtensionDelegate.targets.any { it.platformType == KotlinPlatformType.wasm }
         if (hasWasmTargets) {
             project.logger.warn("WASM targets are added, but UniFFI KMP bindings does not support WASM targets yet.")
         }
 
         val hasAndroidJvmTargets =
-            kotlinMultiplatformExtension.targets.any { it.platformType == KotlinPlatformType.androidJvm }
+            kotlinExtensionDelegate.targets.any { it.platformType == KotlinPlatformType.androidJvm }
         if (hasAndroidJvmTargets && !::androidDelegate.isInitialized) {
             throw GradleException("Android JVM targets are added, but Android Gradle Plugin is not found.")
         }
@@ -239,7 +263,7 @@ class CargoPlugin : Plugin<Project> {
             }
             for (kotlinTarget in cargoBuild.kotlinTargets) {
                 when (kotlinTarget) {
-                    is KotlinJvmTarget -> {
+                    is KotlinJvmTarget, is KotlinWithJavaTarget<*, *> -> {
                         cargoBuild as CargoJvmBuild<*>
                         cargoBuild.variants {
                             configureJvmPostBuildTasks(
@@ -280,7 +304,7 @@ class CargoPlugin : Plugin<Project> {
     }
 
     private fun Project.configureJvmPostBuildTasks(
-        kotlinTarget: KotlinJvmTarget,
+        kotlinTarget: KotlinTarget,
         cargoBuildVariant: CargoJvmBuildVariant<*>,
         androidTarget: KotlinAndroidTarget?,
     ) {
@@ -327,7 +351,12 @@ class CargoPlugin : Plugin<Project> {
                 resources.srcDir(resourceDirectory)
             }
             tasks.withType<ProcessResources> {
-                if (name.contains(kotlinTarget.name)) {
+                @OptIn(InternalGobleyGradleApi::class)
+                if (kotlinExtensionDelegate.pluginId == PluginIds.KOTLIN_JVM) {
+                    if (name == "processResources") {
+                        dependsOn(copyTask)
+                    }
+                } else if (name.contains(kotlinTarget.name)) {
                     dependsOn(copyTask)
                 }
             }