+++ build-logic/gobley-gradle/src/main/kotlin/kotlin/GobleyKotlinAndroidExtensionDelegate.kt
@@ -1,0 +1,40 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+package gobley.gradle.kotlin
+
+import gobley.gradle.InternalGobleyGradleApi
+import gobley.gradle.PluginIds
+import org.gradle.api.DomainObjectCollection
+import org.gradle.api.NamedDomainObjectCollection
+import org.gradle.api.Project
+import org.gradle.kotlin.dsl.getByType
+import org.jetbrains.kotlin.gradle.dsl.KotlinAndroidProjectExtension
+import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
+import org.jetbrains.kotlin.gradle.plugin.KotlinTarget
+
+@OptIn(InternalGobleyGradleApi::class)
+class GobleyKotlinAndroidExtensionDelegate(
+    project: Project
+) : GobleyKotlinExtensionDelegate {
+    private val kotlinAndroidExtension: KotlinAndroidProjectExtension =
+        project.extensions.getByType()
+
+    override val pluginId = PluginIds.KOTLIN_ANDROID
+
+    override val targets: DomainObjectCollection<KotlinTarget> =
+        project.container(KotlinTarget::class.java)
+
+    override val sourceSets: NamedDomainObjectCollection<KotlinSourceSet>
+        get() = kotlinAndroidExtension.sourceSets
+
+    override val implementationVersion: String?
+        get() = kotlinAndroidExtension.javaClass.`package`.implementationVersion
+
+    init {
+        kotlinAndroidExtension.target { targets.add(this) }
+    }
+}